clean up apiHooks index file
