+++ src/apiHooks/library.js
@@ -4,8 +4,12 @@ import { useDispatch } from "react-redux";
 import { Server } from "api";
 import type { LibraryManga } from "@tachiweb/api-client";
 import type { LibraryFlagsType, LibraryFlagsPossibleValueTypes } from "types";
-import { fetcherUnpackData, serialPromiseChain } from "./utils";
 import { useUpdateChapters } from "./chapters";
+import {
+  fetcherUnpackData,
+  fetcherUnpackContent,
+  serialPromiseChain
+} from "./utils";
 
 export function useLibrary() {
   const dispatch = useDispatch();
@@ -82,4 +86,34 @@ export function useSetLibraryFlag(): (
       });
     }
   };
+}
+
+export type UnreadMap = { [mangaId: number]: number };
+
+export function useUnread() {
+  const dispatch = useDispatch();
+
+  return useSWR<UnreadMap>(
+    Server.libraryUnread(),
+    url => fetcherUnpackContent(url).then(content => unreadArrayToMap(content)),
+    {
+      onError(error) {
+        dispatch({
+          type: "library/FETCH_UNREAD_FAILURE",
+          errorMessage: "Failed to get unread chapters for your library",
+          meta: { error }
+        });
+      }
+    }
+  );
+}
+
+function unreadArrayToMap(
+  unreadArray: { id: number, unread: number }[]
+): UnreadMap {
+  const newUnread = {};
+  unreadArray.forEach(unreadObj => {
+    newUnread[unreadObj.id] = unreadObj.unread;
+  });
+  return newUnread;
 }
